Update make-share.sh
On recent Ubuntu, 'time' is provided either by the 'time' package, or as a bash builtin. It's not present in sh, or installed by default. This causes 'pip install keystone-engine' to fail.

So, one of the following needs to happen:
-'time' package is added to the pre-reqs
-switch this script to bash
-remove the 'time' call around make

I've axed the call to 'time', but really whatever you think is best.