Add asm(..., as_bytes=True) to Python bindings
When enabled, the assembled code will be returned as bytes (str on
Python 2) instead of a list of integers, this is more memory efficient,
faster (It's created using ctypes.string_at instead of a loop in
Python), and more convenient since that's the type that most code that
works with bytes in Python expects (That includes capstone).

Defaults to `False` for backwards compatibility, since changing the
return type is a breaking change.