Haskell bindings: fix sdist generation/pkg install
Cabal adds C source files to the sdist automatically, but not the
contents of header dirs. That must be done explicitly. With this
change,`cabal sdist` will produce a buildable sdist (previously, it
would fail due to missing a header file).

Also update package minor version to specifying the fixed version in
dependencies.